backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
redis
celery
django-celery-results
kafka-python
djangorestframework-simplejwt>=5.3.0
channels==4.1.0
channels-redis==4.2.0
daphne>=4.0

razorpay
requests>=2.31.0
django-cors-headers==4.3.1
twilio>=9.0.0



pytest
pytest-django
pytest-cov
